Refactored Frontend. Look at the docs for explanation
