+++ src/common/components/ui/calendar.tsx
@@ -1,9 +1,9 @@
-import * as React from "react"
 import { ChevronLeft, ChevronRight } from "lucide-react"
+import * as React from "react"
 import { DayPicker } from "react-day-picker"
 
+import { buttonVariants } from "@/common/components/ui/button"
 import { cn } from "@/lib/utils"
-import { buttonVariants } from "@/components/ui/button"
 
 function Calendar({
   className,
