+++ src/features/auth/SignUp.jsx
@@ -10,9 +10,9 @@ import { RedSpan } from '@/common/components/form/styles';
 
 import SubmitButton from '@/common/components/form/SubmitButton';
 import GoogleButton from '@/common/components/GoogleButton';
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/common/components/ui/card';
+import { Input } from '@/common/components/ui/input';
+import { Label } from '@/common/components/ui/label';
 
 
 export default function SignUp() {
